+++ robozonky-api/src/main/java/com/github/robozonky/api/remote/entities/SellRequest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 The RoboZonky Project
+ * Copyright 2018 The RoboZonky Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -21,7 +21,7 @@ import javax.xml.bind.annotation.XmlElement;
 
 public class SellRequest extends BaseEntity {
 
-    private int investmentId;
+    private long investmentId;
     private BigDecimal feeAmount, remainingPrincipal;
 
     public SellRequest(final RawInvestment investment) {
@@ -31,7 +31,7 @@ public class SellRequest extends BaseEntity {
     }
 
     @XmlElement
-    public int getInvestmentId() {
+    public long getInvestmentId() {
         return investmentId;
     }
